requirements.txt
beautifulsoup4==4.12.3
greenlet==3.0.3
playwright==1.43.0
pyee==11.1.0
python-dotenv==1.0.1
soupsieve==2.5
typing_extensions==4.11.0
